perf(dashboard): lazy-load recharts performance chart to split bundle
Extract the Real-time Performance History chart into PerformanceChart.tsx
and load it via React.lazy + Suspense with a skeleton fallback. recharts
moves into its own on-demand chunk, shrinking the DashboardPage chunk from
401.64 kB to 17.65 kB (gzip 110.59 -> 4.86 kB). Chart data wiring and
appearance unchanged.
